modules/module2_analysis/requirements.txt
langchain-openai>=0.1.0
langchain-core>=0.1.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0